Add vertebrates component
