+++ src/components/nrc-content/vertebrates/vertebrates-selectors.js
@@ -1,0 +1,7 @@
+import { createStructuredSelector } from 'reselect';
+
+import { getCountryData } from 'selectors/nrc-selectors';
+
+export default createStructuredSelector({
+  countryData: getCountryData,
+});
