+++ src/components/nrc-content/vertebrates/vertebrates-styles.module.scss
@@ -1,0 +1,80 @@
+@import 'styles/settings';
+@import 'styles/typography-extends';
+
+
+
+.vertebratesContainer {
+  background: $dark;
+  border: 1px solid $elephant;
+  border-radius: 8px;
+  color: $white;
+  display: flex;
+  justify-content: space-between;
+  margin: 8px 40px 8px 20px;
+  padding: 10px 16px;
+
+  .endemicCardsContainer {
+    @extend %bodyText;
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+
+    .endemicCard {
+      align-items: center;
+      display: flex;
+      margin-right: 15px;
+    }
+
+    .endemicIcon {
+      background-color: $white;
+      border-radius: 50%;
+      padding: 1px;
+      height: 32px;
+      width: 32px;
+      margin-right: 8px;
+    }
+  }
+  p {
+    margin: 0;
+  }
+
+  button {
+    border-radius: $border-radius-desktop;
+    margin-left: 20px;
+    padding: 10px 32px;
+    white-space: nowrap;
+  }
+}
+
+.titleTooltip {
+  @extend %annotation;
+  border-radius: $border-radius-desktop;
+  background: $white;
+  box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.05);
+  color: $dark-text;
+  margin-bottom: 5px;
+  padding: 2px 6px;
+  width: 200px;
+}
+
+.titleTooltip::after {
+  content: "";
+  position: absolute;
+  top: 100%;
+  left: 50%;
+  margin-left: -5px;
+  margin-top: -5px;
+  border-width: 5px;
+  border-style: solid;
+  border-color: white transparent transparent transparent;
+}
+
+.nrcContentVertebrates:before {
+  content: '';
+  position: absolute;
+  bottom: 60px;
+  width: 100%;
+  height: 50px;
+  z-index: 9;
+  background-image: linear-gradient(to top, $app-background 0%, rgba(9, 12, 20, 0) 100%);
+}
